backend/requirements.txt
asgiref==3.8.1
click==8.2.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.3.0
pillow==11.2.1
sqlparse==0.5.3
